Add the web import workaround
