+++ packages/path_provider/path_provider_foundation/example/integration_test/path_provider_test.dart
@@ -9,8 +9,8 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:objective_c/objective_c.dart';
 import 'package:path/path.dart' as p;
-import 'package:path_provider_foundation/path_provider_foundation.dart';
 import 'package:path_provider_foundation/src/ffi_bindings.dart';
+import 'package:path_provider_foundation/src/path_provider_foundation_real.dart';
 import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
 
 import 'path_provider_test.mocks.dart';
